Reorganized kafka connector crate.
